Added Hit Stop controller
A single function that does the hit stop. It turns Time.timescale to zero for a specified duration. After that duration it sets Time.timescale back to whatever value it was prior to the function being called (that way it doesn't override any other timescale affects). Doing a hit stop in the middle of hit stop does nothing.

Also modified CamShakeController slightly so that it does nothing if duration is zero or less, and so it doesn't shake again if it's already shaking.

Also simply renamed TimerTestScript to ScriptTester cause I started jamming other tests in there.